Start of automated trading python script
Right now checks for a timestamp and prints AMD quote every 10 seconds

Goal of v1: Check quote when market is open(so 6:30 AM) and store. Then check quote every 10 seconds or so. If price is 3% or so higher than when market opens, buy. Record this price. If market dips 3% or below, then sell

So far, code is always running through a while true loop - make this more robust. Only run if market is open 6:30 AM to 1PM or so

So far this has been tested on windows/spyder. Test on Odroid, and upload to Initial State web api for text logging and plotting purposes. Still need to install Initial State ISSstreamer on odroid