Got test report generated, and email forwarding working
Got text logging to report_generation.txt - more to add this in coming days. Buy sell function still needs lots of work. Get that working first with buy/sell limits, and then think about making a stock object instead of the array structure the program currently uses - it's inefficient and won't scale.

Email forwarding works with attachments - Once v1 of the report is ready to go, make a loop in email_reporting.py so that it emails updates every 30 minutes or so